% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rppl.R
\name{augment_data}
\alias{augment_data}
\title{Augments a data frame with dummy variables}
\usage{
augment_data(data, l)
}
\arguments{
\item{data}{data frame}

\item{l}{a list of vectors to augment with}
}
\value{
data frame
}
\description{
Augments a data frame with dummy variables
}
