% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rppl.R
\name{impose_sort_order}
\alias{impose_sort_order}
\title{Impose sort order}
\usage{
impose_sort_order(data, params)
}
\arguments{
\item{data}{a data frame that should be sorted}

\item{params}{a list of params that have to be checked}
}
\value{
TRUE if no problem was found, false otherwise
}
\description{
TODO check if needed
}
