% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rppl.R
\name{json_processor}
\alias{json_processor}
\title{Processes the JSON simulation output by npm run sim}
\usage{
json_processor(sim, npar, reps, name = "R")
}
\arguments{
\item{sim}{simulation object (JSON)}

\item{npar}{the number of parameters that simulation had (tree is counted)}

\item{name}{string to be used for the values}
}
\value{
a data frame
}
\description{
Processes the JSON simulation output by npm run sim
}
