% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rppl.R
\name{merge_models}
\alias{merge_models}
\title{Merge two models}
\usage{
merge_models(data_m1, data_m2, augment)
}
\arguments{
\item{augment}{values used to augment model1 to be mergable with model}
}
\value{
merged augmented models
}
\description{
Merge two models
}
