% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rppl.R
\name{numpar_model}
\alias{numpar_model}
\title{Get the number of parameters of model}
\usage{
numpar_model(modelf)
}
\arguments{
\item{modelf}{model file}
}
\value{
the number of parameters
}
\description{
Get the number of parameters of model
}
