% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rppl.R
\name{read_model_description}
\alias{read_model_description}
\title{Read Verbose Model Description}
\usage{
read_model_description(modelf)
}
\arguments{
\item{modelf}{model file}
}
\value{
character
}
\description{
Read Verbose Model Description
}
