% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rppl.R
\name{strip_read_phyjson}
\alias{strip_read_phyjson}
\title{Strip PhyJSON Reading Function}
\usage{
strip_read_phyjson(x)
}
\arguments{
\item{x}{character(1)  A JS string for reading a PhyJSON tree.}
}
\value{
character(1)   The pure tree name.
}
\description{
Converts strings such as
}
\details{
phyjs.read_phyjson('../../data/bird-phyjson/Accipitridae.MCC.tre.phyjson')

to simply the tree name, e.g. to 'Accipitridae.MCC.tre'.
}
\examples{

strip_read_phyjson(x = "phyjs.read_phyjson(../../data/bird-phyjson/Accipitridae.MCC.tre.phyjson')")
}
