% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rppl.R
\name{sumstat_var}
\alias{sumstat_var}
\title{Compute the summary statistic variance}
\usage{
sumstat_var(data, npar, reps)
}
\arguments{
\item{data}{data frame}

\item{npar}{number of parameter columns}

\item{reps}{number of replicates}
}
\value{
a vector of the row means of the replicates
}
\description{
Compute the summary statistic variance
}
