% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rppl.R
\name{test_summary}
\alias{test_summary}
\title{Summarizes unit test results}
\usage{
test_summary(r)
}
\arguments{
\item{r}{data frame that has the unit test results (must have a column called PASS)}
}
\value{
a list of the summaries
}
\description{
Summarizes unit test results
}
