% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rppl.R
\name{update_db}
\alias{update_db}
\title{Updates database}
\usage{
update_db(dataf, newjs, modelname, reps)
}
\arguments{
\item{dataf}{where the database is stored}

\item{newjs}{new data in json format}

\item{modelname}{the name of the model}

\item{reps}{the number of reps}
}
\value{
data frame with updated db
}
\description{
TODO modelname and reps should probably be parsed automatically
}
